+++ djangocms_oidc_form_fields/tests/test_cms_plugins.py
@@ -1,7 +1,12 @@
 from aldryn_forms.models import FormPlugin
 from cms.api import add_plugin, create_page
+from cms.models import Placeholder
+from cms.plugin_rendering import ContentRenderer
 from cms.test_utils.testcases import CMSTestCase
+from django.test import RequestFactory
 from djangocms_oidc.constants import DJANGOCMS_USER_SESSION_KEY
+
+from djangocms_oidc_form_fields.cms_plugins import OIDCSpanElement
 
 
 class TestOIDCFormPlugin(CMSTestCase):
@@ -100,3 +105,51 @@ class TestOIDCFormPlugin(CMSTestCase):
             "email": "tester@foo.foo"
         })
         self.assertEqual(response.context['form'].errors, {'email': ['This field is required.']})
+
+
+class TestOIDCElementPlugin(CMSTestCase):
+
+    def _create_model(self, plugin_class, **kwargs):
+        placeholder = Placeholder.objects.create(slot='test')
+        model_instance = add_plugin(
+            placeholder,
+            plugin_class,
+            'en',
+            **kwargs
+        )
+        return model_instance
+
+    def _create_request(self, data=None):
+        request = RequestFactory().request()
+        request.session = {}
+        if data is not None:
+            request.session[DJANGOCMS_USER_SESSION_KEY] = data
+        return request
+
+    def test_no_attrs(self):
+        model_instance = self._create_model(OIDCSpanElement, oidc_attributes="given_name")
+        request = self._create_request()
+        renderer = ContentRenderer(request=request)
+        html = renderer.render_plugin(model_instance, {'request': request})
+        self.assertHTMLEqual(html, """<span class="oidc-data given_name"></span>""")
+
+    def test_with_attrs(self):
+        model_instance = self._create_model(OIDCSpanElement, oidc_attributes="given_name")
+        request = self._create_request({"given_name": "The Tester"})
+        renderer = ContentRenderer(request=request)
+        html = renderer.render_plugin(model_instance, {'request': request})
+        self.assertHTMLEqual(html, """<span class="oidc-data given_name">The Tester</span>""")
+
+    def test_with_formatted_attrs(self):
+        model_instance = self._create_model(OIDCSpanElement, oidc_attributes="address")
+        request = self._create_request({'address': {'formatted': 'Street 42\n123 00 City'}})
+        renderer = ContentRenderer(request=request)
+        html = renderer.render_plugin(model_instance, {'request': request})
+        self.assertHTMLEqual(html, """<span class="oidc-data address">Street 42\n123 00 City</span>""")
+
+    def test_with_multilevel_attrs(self):
+        model_instance = self._create_model(OIDCSpanElement, oidc_attributes="address.location.city")
+        request = self._create_request({"address": {"location": {"city": "Prague"}}})
+        renderer = ContentRenderer(request=request)
+        html = renderer.render_plugin(model_instance, {'request': request})
+        self.assertHTMLEqual(html, """<span class="oidc-data addresslocationcity">Prague</span>""")